Move PointerMutability from expr to ty module
